requirements.txt
smartapi-python
logzero
pyotp
websocket
fastapi
uvicorn
yfinance
pandas
numpy 
nsepy
httpx
requests
bs4 

